update the code style with helper function
