+++ src/client/utils/api.ts
@@ -1,5 +1,7 @@
 import axios from "axios";
 
+import { saveFile } from "@/client/utils/helper";
+
 export const uploadFiles = async (files: FileList) => {
   const data = new FormData();
   for (let i = 0; i < files.length; i++) {
@@ -14,13 +16,5 @@ export const uploadFiles = async (files: FileList) => {
 
 export const downloadFile = async (name: string) => {
   const res = await axios.get(`/api/file/download/${name}`, { responseType: "blob" });
-  const blob = new Blob([res.data]);
-  const url = URL.createObjectURL(blob);
-  const a = document.createElement('a');
-  a.href = url;
-  a.download = name;
-  document.body.appendChild(a);
-  a.click();
-  URL.revokeObjectURL(url);
-  document.body.removeChild(a);
+  saveFile(name, new Blob([res.data]));
 }
